Add symbol handling and cleanup, including renaming.
Renamed from gtIrbPlugin to gtirb, the fact that is a plugin is
given, and the Plugin prefix is repeatedly in the automatic naming
of the generated files.
This version imports symbols OK though much room for improvement.
Also included in this checkin is a zip file that can be added to
any Ghidra 9.1 installation to add this plugin.