Added ability to export GTIRB from other sources
Before this change, the GTIRB exporter could only be used with Ghidra
projects that were loaded from the GTIRB loader as most exported data
was copied from that source file. Now, it will export the information
from Ghidra's program data. The source GTIRB is still used if one exists
in order to  preserve UUIDs and AuxData schemas that this exporter does
not (yet) support populating.